requirements.txt
numpy==1.19.5
matplotlib==3.5.0
opencv_python==4.5.4.60
tensorflow_gpu==2.4.0
Pillow==8.4.0
pandas==1.3.4
sklearn